Tengen: Move chatWidget to gui library.
