+++ src/game/gui/CMakeLists.txt
@@ -5,11 +5,13 @@ find_package(Qt6 REQUIRED COMPONENTS Widgets)
 # Get files to build
 set(headers
     "${CMAKE_CURRENT_LIST_DIR}/include/gui/boardWidget.hpp"
+    "${CMAKE_CURRENT_LIST_DIR}/include/gui/chatWidget.hpp"
     "${CMAKE_CURRENT_LIST_DIR}/boardRenderer.hpp"
 )
 
 set(sources
     "${CMAKE_CURRENT_LIST_DIR}/boardWidget.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/chatWidget.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/boardRenderer.cpp"
 )
